- Update utils functions from UIExtension.kt to Utils.
